Grid2D, GridCell
A Cartesian subdivision of an integer-valued input domain into approximately evenly-sized subrectangles. Similar to the implementation in C++ and C#.

Currently does not implement the "aligned" initialization mode because MATLAB does not have a way to take advantage of it. Refer to OrdinalDownsamp for client code modifications needed to use aligned mode.