AMC, CMC PresenceMask
CommonMovementCache now maintains a presence indicator matrix that allows quick lookup of a particular run of CommonMovementClassifier that matches a particular set of image IDs.

This change requires CommonMovementCache to know the complete set of image IDs (including those that are never ever processed). AdaptiveMovementClassifier has been modified to provide this list of image IDs to the ctor of CommonMovementCache.