CommonMovementCache added Keys
CommonMovementCache.Keys is a column cell vector, each cell containing the image IDs consumed in a particular run of CommonMovementClassifier.

This property provides the same information as the PresenceMark property, but in a different representation.

Note: In the rare usage case where CommonMovementClassifier is constructed from an image id list that is NOT in ascending order, the Keys property will preserve the ordering used to initialize CommonMovementClassifier, whereas the PresenceMask will not.