MovementGraph initial commit
MovementGraph creates a connectivity graph based on movement of image content across a set of images.

The inputs used to compute the MovementGraph comes from CommonMovementCache, which contains instances of CommonMovementClassifier, each of which is in turn computed from small subsets of images having shared image content shifted across the screen.